Create project sprint-1/step-1 with some fixes
