+++ package.json
@@ -11,6 +11,7 @@
     "@types/react": "^18.0.4",
     "@types/react-dom": "^18.0.0",
     "@ya.praktikum/react-developer-burger-ui-components": "^1.13.1",
+    "prop-types": "^15.8.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-scripts": "5.0.1",
