+++ public/index.html
@@ -8,8 +8,7 @@
       name="description"
       content="Web site created using create-react-app"
     />
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <title>React App</title>
+    <title>CosmoReactBurger</title>
   </head>
   <body>
     <div id="root"></div>
