--- src/Components/App/app.module.css
@@ -1,30 +1,0 @@
-.myFlex{
-    display: flex;
-    flex-flow: row nowrap;
-    justify-content: center;
-    align-items: center;
-}
-
-.main{
-    composes:myFlex;
-    align-items: flex-start;
-    height:90vh;
-
-
-}
-
-
-
-.app_text{
-    margin: 0 0 0 1vw;
-}
-
-@media screen and (max-width: 1400px){
-    .main{
-        flex-flow: column nowrap;
-        height:fit-content;
-        row-gap: 5vw;
-        align-items: center;
-    }
-
-}
