--- src/Components/App/appStyles.css
@@ -1,3 +1,0 @@
-.app_text{
-    margin:0 0 0 2vw;
-}
