--- src/Components/burgerConstructor/burgerConstructor.module.css
@@ -1,78 +1,0 @@
-.myFlex{
-    display: flex;
-    flex-flow: row nowrap;
-    justify-content: center;
-    align-items: center;
-}
-
-.area{
-    composes: myFlex;
-    width:40vw;
-    height:90vh;
-    flex-flow: column nowrap;
-    row-gap: 10px;
-}
-
-.price{
-    composes: myFlex;
-    width: 40vw;
-    height: 10vh;
-    justify-content: flex-end;
-    column-gap: 1vw;
-}
-
-.order{
-    composes: myFlex;
-    flex-flow: column nowrap;
-    width: 40vw;
-    height: 80vh;
-    row-gap:10px;
-}
-
-.middle{
-    composes: myFlex;
-    flex-flow: column nowrap;
-    justify-content: flex-start;
-    width: 30vw;
-    height: 60vh;
-    overflow: auto;
-    row-gap:10px;
-    scrollbar-width: thin;
-    scrollbar-color: #8484ac #2f2f37;
-}
-
-.position{
-    composes: myFlex;
-    width:90%;
-    column-gap:20px;
-}
-
-
-
-@media screen and (max-width: 1400px){
-
-    .middle {
-        width: 98vw;
-    }
-
-    .area{
-        width:98vw;
-
-    }
-}
-
-@media screen and (max-width: 600px){
-    .middle {
-        width: 90vw;
-    }
-    .area{
-        width:90vw;
-
-    }
-    .constructor-element {
-        padding:0;
-        min-width: 95%;
-    }
-
-
-}
