+++ src/Components/burgerIngridients/ingredientList/ingredient-list.module.css
@@ -6,8 +6,10 @@
 }
 .cardArea{
     composes: myFlex;
-    justify-content: flex-start;
     flex-flow: row wrap;
+    margin: 24px 0 0 16px;
+    column-gap: 24px;
+    row-gap: 32px;
 }
 .card{
     display: flex;
@@ -15,13 +17,16 @@
     position:relative;
     justify-content: center;
     align-items: center;
-    width:10vw;
+    width:fit-content;
     height:fit-content;
-    margin:1vw;
-    padding: 1vw;
-    border-radius:10px;
-    row-gap:10px;
+    row-gap:4px;
+    max-width: 272px;
 
+}
+
+
+.cardImage{
+    margin: 0 16px 0 16px;
 }
 
 @media screen and (max-width: 1400px) and (min-width:1000px){
