+++ src/Components/header/header.module.css
@@ -1,3 +1,10 @@
+/*базовый сброс отступов*/
+*{
+    margin: 0;
+    padding: 0;
+}
+
+/*общий шаблон для стилей flex-контейнеров*/
 .flexBlock{
     display:flex;
     flex-flow: row nowrap;
@@ -5,45 +12,70 @@
     align-items: center;
 }
 
-.header{
+/*общее поле заголовка*/
+.top{
     composes: flexBlock;
-    background: #131316;
-    padding: 10px;
-    column-gap: 2vw;
-
+    background: #1c1c21;
 }
 
+/*контейнер с содержимым*/
+.header{
+    composes: flexBlock;
+    padding: 10px;
+    column-gap: 2vw;
+    width: 1280px;
+}
+
+
+/*зона левых кнопок*/
+.btnArea{
+    composes: flexBlock;
+    column-gap:8px;
+    margin: 0 0 0 16px;
+    flex: 1 1 auto;
+    justify-content: flex-end;
+    height:fit-content;
+}
+
+/*зона кнопок конструктора*/
 .constructor{
+    composes: flexBlock;
+    justify-content: flex-end;
+    flex: 1 1 auto;
+    column-gap: 8px;
+    height:fit-content;
+}
+
+/*зона кнопок заказа*/
+.orderList{
+    composes: flexBlock;
+    justify-content: flex-start;
+    flex: 1 1 auto;
+    column-gap: 8px;
+    padding:0 0 0 20px;
+    height:fit-content;
+}
+
+/*логотип*/
+.logo{
+    composes: flexBlock;
+    flex: 1 1 auto;
+    margin: 16px 0 0 16px;
+}
+
+/*зона перс. аккаунта*/
+.personalAccount{
     composes: flexBlock;
     justify-content: flex-end;
     column-gap:10px;
     flex: 1 1 auto;
+    margin: 0 20px 0 16px;
 }
 
-.personalAccount{
-    composes: flexBlock;
-    justify-content: flex-start;
-    column-gap:10px;
-    flex: 1 1 auto;
-}
 
-.logo{
-    composes: flexBlock;
-    flex: 1 1 auto;
-}
 
-.orderList{
-    composes: flexBlock;
-    justify-content: flex-start;
-    margin: 0 0 0 3vw;
-    column-gap:10px;
-    flex: 1 1 auto;
+@media screen and (max-width: 600px){
+    .headerText{
+        display:none;
+    }
 }
-
-.btnArea{
-    composes: flexBlock;
-    column-gap:10px;
-    flex: 2 2 auto;
-    justify-content: center;
-}
-
