--- src/Components/header/styles.css
@@ -1,5 +1,0 @@
-@media screen and (max-width: 600px){
-    .headerText{
-        display:none;
-    }
-}
